HGF

Documentation for ActionModels.

HGF

Tutorials

Functions

Tutorials

Functions

ActionModels.premade_agents

ActionModels.Agent

ActionModels.RejectParameters

ActionModels.fit_model

ActionModels.get_history

ActionModels.get_history

ActionModels.get_history

ActionModels.get_history

ActionModels.get_history

ActionModels.get_posteriors

ActionModels.get_states

ActionModels.get_states

ActionModels.get_states

ActionModels.get_states

ActionModels.give_inputs!

ActionModels.give_inputs!

ActionModels.give_inputs!

ActionModels.init_agent

ActionModels.init_agent

ActionModels.multiple_actions

ActionModels.plot_predictive_simulation

ActionModels.plot_trajectory

ActionModels.plot_trajectory

ActionModels.plot_trajectory!

ActionModels.plot_trajectory!

ActionModels.premade_agent

ActionModels.reset!

ActionModels.reset!

ActionModels.set_params!

ActionModels.set_params!

ActionModels.set_params!

ActionModels.single_input!

MCMCChains.get_params

MCMCChains.get_params

MCMCChains.get_params

MCMCChains.get_params

ActionModels.premade_agents — Constant

ActionModels.Agent — Type

ActionModels.RejectParameters — Type

Custom error type which will result in rejection of a sample

ActionModels.fit_model — Function

"" function fitmodel( agent::Agent, inputs::Array, actions::Vector, parampriors::Dict, fixedparams::Dict = Dict(); sampler = NUTS(), niterations = 1000, n_chains = 1, verbose = true, ) Function to fit an agent's parameters.

ActionModels.get_history — Method

ActionModels.get_history — Method

ActionModels.get_history — Method

ActionModels.get_history — Method

ActionModels.get_history — Method

ActionModels.get_posteriors — Method

ActionModels.get_states — Method

ActionModels.get_states — Method

ActionModels.get_states — Method

ActionModels.get_states — Method

ActionModels.give_inputs! — Method

give_inputs!(agent::Agent, inputs::Array)

Function for inputting multiple observations to an agent. Input is structured as an Array, with each row being a single input consisting of multiple numbers.

ActionModels.give_inputs! — Method

give_inputs!(agent::Agent, inputs::Real)

Convenience method for inputting multiple observations to an agent. Input is here just a single value.

ActionModels.give_inputs! — Method

give_inputs!(agent::Agent, inputs::Vector)

Function for inputting multiple observations to an agent. Input is structured as an Array, with each row being a single input consisting of multiple numbers.

ActionModels.init_agent — Method

ActionModels.init_agent — Method

ActionModels.multiple_actions — Method

ActionModels.plot_predictive_simulation — Method

ActionModels.plot_trajectory! — Method

ActionModels.plot_trajectory! — Method

ActionModels.plot_trajectory — Method

ActionModels.plot_trajectory — Method

ActionModels.premade_agent — Function

function premade_agent(
    model_name::String, params_list::NamedTuple = (;)
)

Function for making a premade agent.

ActionModels.reset! — Method

ActionModels.reset! — Method

ActionModels.set_params! — Method

ActionModels.set_params! — Method

ActionModels.set_params! — Method

ActionModels.single_input! — Method

single_input!(agent::Agent, input)

Function for giving an input to an Agent.

MCMCChains.get_params — Method

MCMCChains.get_params — Method

MCMCChains.get_params — Method

MCMCChains.get_params — Method